requirements.txt
fpdf
matplotlib
mysql-connector-python
python-dotenv

